refactor: mock 정리 (#129)
- 샘플용 mock 데이터 정의
- 테스트 타겟 mock 객체 폴더 분리
