+++ Molio/Source/Domain/Entity/MolioPlaylist.swift
@@ -17,3 +17,13 @@ struct MolioPlaylist {
         )
     }
 }
+
+extension MolioPlaylist {
+    static let mock = MolioPlaylist(
+        id: UUID(uuidString: "12345678-1234-1234-1234-1234567890ab")!,
+        name: "목플레이리스트",
+        createdAt: Date(),
+        musicISRCs: [],
+        filter: MusicFilter(genres: [.pop, .acoustic])
+    )
+}
