+++ MolioTests/PublishAllMusicInCurrentPlaylistUseCaseTests.swift
@@ -8,14 +8,14 @@ final class PublishAllMusicInCurrentPlaylistUseCaseTests: XCTestCase {
         name: "My Favorite Songs",
         createdAt: Date(),
         musicISRCs: ["USUM72014967", "USUM72014968"],
-        filters: ["Pop", "2020s"]
+        filter: MusicFilter(genres: [.pop, .metal])
     )
     private let newPlaylist = MolioPlaylist(
         id: UUID(),
         name: "Chill Vibes",
         createdAt: Date(),
         musicISRCs: ["USUM72014969", "USUM72014970"],
-        filters: ["Chill", "Acoustic"]
+        filter: MusicFilter(genres: [.chill, .acoustic])
     )
 
     private var subscriptions = Set<AnyCancellable>()
@@ -56,33 +56,5 @@ final class PublishAllMusicInCurrentPlaylistUseCaseTests: XCTestCase {
         
         // Now perform your assertions
         XCTAssertNotEqual(emittedMusicISRCs[0], emittedMusicISRCs[1], "First and second playlists should have different music ISRCs.")
-    }
-}
-
-
-// Mock 구현을 위한 클래스
-private class MockPublishCurrentPlaylistUseCase: PublishCurrentPlaylistUseCase {
-    
-    // 원하는 MolioPlaylist을 설정할 수 있는 프로퍼티
-    var mockPlaylist: MolioPlaylist?
-    
-    // 데이터를 발행하기 위한 퍼블리셔
-    private var subject: CurrentValueSubject<MolioPlaylist?, Never>
-    
-    // 초기화 시 mockPlaylist를 설정하고 subject를 초기화
-    init(mockPlaylist: MolioPlaylist? = nil) {
-        self.mockPlaylist = mockPlaylist
-        self.subject = CurrentValueSubject<MolioPlaylist?, Never>(mockPlaylist)
-    }
-    
-    // 프로토콜의 execute 메서드 구현
-    func execute() -> AnyPublisher<MolioPlaylist?, Never> {
-        return subject.eraseToAnyPublisher()
-    }
-    
-    // 테스트 중에 mockPlaylist를 업데이트할 수 있는 메서드
-    func updatePlaylist(_ playlist: MolioPlaylist?) {
-        self.mockPlaylist = playlist
-        subject.send(playlist)
     }
 }
