+++ MolioTests/PublishCurrentPlaylistUseCaseTests.swift
@@ -60,57 +60,3 @@ final class PublishCurrentPlaylistUseCaseTests: XCTestCase {
         }
     }
 }
-
-
-
-private class MockCurrentPlaylistRepository: CurrentPlaylistRepository {
-    func setDefaultPlaylist(_ id: UUID) throws {
-        
-    }
-    
-    private var currentPlaylistUUID = CurrentValueSubject<UUID?, Never>(nil)
-    
-    var currentPlaylistPublisher: AnyPublisher<UUID?, Never> {
-        currentPlaylistUUID.eraseToAnyPublisher()
-    }
-    
-    func setCurrentPlaylist(_ id: UUID) {
-        currentPlaylistUUID.send(id)
-    }
-}
-
-private class MockPlaylistRepository: PlaylistRepository {
-    private var playlists = CurrentValueSubject<[MolioPlaylist], Never>([])
-    
-    var playlistsPublisher: AnyPublisher<[MolioPlaylist], Never> {
-        playlists.eraseToAnyPublisher()
-    }
-    
-    func addMusic(isrc: String, to playlistName: String) {
-    }
-    
-    func deleteMusic(isrc: String, in playlistName: String) {
-
-    }
-    
-    func deletePlaylist(_ playlistName: String) {
-        
-    }
-    
-    func fetchPlaylist(for name: String) -> MolioPlaylist? {
-        MolioPlaylist(id: UUID(), name: name, createdAt: Date.now, musicISRCs: [], filters: [])
-    }
-    
-    func fetchPlaylists() -> [MolioPlaylist]? {
-        []
-    }
-    
-    func moveMusic(isrc: String, in playlistName: String, fromIndex: Int, toIndex: Int) {
-        
-    }
-    
-    func saveNewPlaylist(_ playlistName: String) async throws -> UUID {
-        return UUID()
-    }
-
-}
